Fixed a pebkac error
Textboxes are drawn in a container that's above everything else and I had the opacity set to 255. If a meter is drawn in the container, the container isn't drawn. But if a game doesn't have a textbox then the container is drawn and covers the whole screen.

Also updated the older games to use the new scale variable.